Move CreateTeam to CreateGroup
