+++ src/index.js
@@ -11,7 +11,7 @@ import Layout from "./app/components/layout";
 import SignInController from "./app/views/login/SignIn";
 import ForgottenPasswordController from "./app/views/new-password/forgottenPasswordController";
 import Collections from "./app/views/collections";
-import CreateTeam from "./app/views/teams/team-create/CreateTeam"
+import CreateTeam from "./app/views/groups/create/CreateGroup"
 import SelectADataset from "./app/views/datasets-new/DatasetsController";
 import DatasetEditionsController from "./app/views/datasets-new/editions/DatasetEditionsController";
 import DatasetVersionsController from "./app/views/datasets-new/versions/DatasetVersionsController";
